Fixes losing the focus on textarea in Chrome
After a bit of investigation 🧐 I found out that the focus was being
lost when the textarea is disabled. Only Chrome has this behavior afaik.

The proposed solution avoids changing the enabled/disabled status
of the textarea and instead has a guard in the `onKeyPress` callback that
will prevent re-sending the message if another one is being sent
already.